refactor: move errors to own module
