+++ src/lib.rs
@@ -1,1 +1,8 @@
+use connection::SnapdConnectionManager;
+use deadpool::managed::Pool;
+
 mod connection;
+
+pub struct SnapdClient {
+    pool: Pool<SnapdConnectionManager>,
+}
